# Conflicts:
#	src/main/java/org/nurim/nurim/Controller/CommunityController.java
#	src/main/java/org/nurim/nurim/Controller/CommunityImageController.java
#	src/main/java/org/nurim/nurim/config/auth/TokenProvider.java
#	src/main/java/org/nurim/nurim/config/auth/TokenValidateFilter.java
#	src/main/java/org/nurim/nurim/service/CommunityImageService.java
#	src/main/java/org/nurim/nurim/service/CommunityService.java
#	src/main/java/org/nurim/nurim/service/ReplyService.java

+++ src/main/java/org/nurim/nurim/service/MemberImageService.java
@@ -54,15 +54,24 @@ public class MemberImageService {
         if (memberImageOptional.isPresent()) {
             MemberImage memberImage = memberImageOptional.get();
 
+            // 이미지의 url과 key가 default와 일치하는지 확인
+            String imageUrl = memberImage.getMemberProfileImage();
+            String imageKey = memberImage.getProfileName();
+
+            // 기본 이미지로 변경할 이미지 경로
+            String defaultImageUrl = "https://nurimplus.s3.ap-northeast-2.amazonaws.com/images/8383f351-73fc-47c5-bf2f-b6ebc105326a.jpeg";
+            String defaultKey = "images/8383f351-73fc-47c5-bf2f-b6ebc105326a.jpeg";
+
+            if (imageUrl.equals(defaultImageUrl) && imageKey.equals(defaultKey)) {
+                // 이미지의 url과 key가 default와 일치하면 삭제 로직을 적용하지 않음
+                return true;
+            }
+
             // S3에서 파일 삭제
             boolean isRemovedFromS3 = fileUploadService.deleteFile(memberImage.getProfileName());
 
             // S3에서 이미지 삭제에 성공하면 DB에서 기본 이미지로 변경
             if (isRemovedFromS3) {
-                // 기본 이미지로 변경할 이미지 경로
-                String defaultImageUrl = "https://nurimplus.s3.ap-northeast-2.amazonaws.com/images/c4e11d02-3ed4-4475-9a57-18918721d381.jpeg";
-                String defaultKey = "images/c4e11d02-3ed4-4475-9a57-18918721d381.jpeg";
-
                 // DB에서 해당 회원의 이미지를 기본 이미지로 변경
                 memberImage.setMemberProfileImage(defaultImageUrl);
                 memberImage.setProfileName(defaultKey);
